+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -128,6 +128,7 @@ CHANGELOG
        }
    }
    ```
+ * Deprecate XML configuration format, use YAML or attributes instead
 
 7.3
 ---